web: extract web hover overlay story into a separate file
