+++ client/web/src/components/WebHoverOverlay/index.ts
@@ -1,0 +1,1 @@
+export * from './WebHoverOverlay'
