+++ client/web/src/components/shared.tsx
@@ -1,65 +1,16 @@
 import classNames from 'classnames'
-import React, { useCallback, useEffect } from 'react'
+import React from 'react'
 
 import { ActionsNavItems, ActionsNavItemsProps } from '@sourcegraph/shared/src/actions/ActionsNavItems'
 import {
     CommandListPopoverButton,
     CommandListPopoverButtonProps,
 } from '@sourcegraph/shared/src/commandPalette/CommandList'
-import { HoverOverlay, HoverOverlayProps } from '@sourcegraph/shared/src/hover/HoverOverlay'
-import { isErrorLike } from '@sourcegraph/shared/src/util/errors'
-import { useLocalStorage } from '@sourcegraph/shared/src/util/useLocalStorage'
 import { useRedesignToggle } from '@sourcegraph/shared/src/util/useRedesignToggle'
-
-import { HoverThresholdProps } from '../repo/RepoContainer'
 
 // Components from shared with web-styling class names applied
 
-export const WebHoverOverlay: React.FunctionComponent<HoverOverlayProps & HoverThresholdProps> = props => {
-    const [dismissedAlerts, setDismissedAlerts] = useLocalStorage<string[]>('WebHoverOverlay.dismissedAlerts', [])
-    const onAlertDismissed = useCallback(
-        (alertType: string) => {
-            if (!dismissedAlerts.includes(alertType)) {
-                setDismissedAlerts([...dismissedAlerts, alertType])
-            }
-        },
-        [dismissedAlerts, setDismissedAlerts]
-    )
-
-    let propsToUse = props
-    if (props.hoverOrError && props.hoverOrError !== 'loading' && !isErrorLike(props.hoverOrError)) {
-        const filteredAlerts = (props.hoverOrError?.alerts || []).filter(
-            alert => !alert.type || !dismissedAlerts.includes(alert.type)
-        )
-        propsToUse = { ...props, hoverOrError: { ...props.hoverOrError, alerts: filteredAlerts } }
-    }
-
-    const { hoverOrError } = propsToUse
-    const { onHoverShown, hoveredToken } = props
-
-    /** Whether the hover has actual content (that provides value to the user) */
-    const hoverHasValue = hoverOrError !== 'loading' && !isErrorLike(hoverOrError) && !!hoverOrError?.contents?.length
-
-    useEffect(() => {
-        if (hoverHasValue) {
-            onHoverShown?.()
-        }
-    }, [hoveredToken?.filePath, hoveredToken?.line, hoveredToken?.character, onHoverShown, hoverHasValue])
-
-    return (
-        <HoverOverlay
-            {...propsToUse}
-            className="card"
-            iconClassName="icon-inline"
-            iconButtonClassName="btn btn-icon"
-            actionItemClassName="btn btn-secondary"
-            infoAlertClassName="alert alert-secondary" // #18931
-            errorAlertClassName="alert alert-danger"
-            onAlertDismissed={onAlertDismissed}
-        />
-    )
-}
-WebHoverOverlay.displayName = 'WebHoverOverlay'
+export { WebHoverOverlay } from './WebHoverOverlay'
 
 export const WebCommandListPopoverButton: React.FunctionComponent<CommandListPopoverButtonProps> = props => {
     const [isRedesignEnabled] = useRedesignToggle()
